+++ lib/dlms-security/include/dlms/security/in_memory_key_store.hpp
@@ -7,7 +7,7 @@
 namespace dlms {
 namespace security {
 
-class InMemoryKeyStore : public IKeyStore
+class InMemoryKeyStore : public IMutableKeyStore
 {
 public:
   SecurityStatus SetKey(const SecurityKey& key);